+++ src/list_index_of.c
@@ -7,9 +7,9 @@
 
 #include "list.h"
 
-int list_index_of(list_t const *list, void const *elem, list_equal_cmp_t cmp)
+long list_index_of(list_t const *list, void const *elem, list_equal_cmp_t cmp)
 {
-    int i = 0;
+    long i = 0;
     list_elem_t *e = list->elems_head;
 
     while (NULL != e) {
