+++ tests/tests_list_index_of.c
@@ -25,10 +25,10 @@ Test(list_index_of, test_impl)
     cr_assert(eq(int, 0, list_insert_at(list, "This is my tail 2-1 !", -1)));
     cr_assert(eq(int, 0, list_insert_at(list, "This is my tail 2 !", 2)));
     cr_assert(eq(int, 4, list_count(list)));
-    cr_assert(eq(int, 0, list_index_of(list, "This is my head !", my_strcmp)));
-    cr_assert(eq(int, 1, list_index_of(list, "This is my tail !", my_strcmp)));
-    cr_assert(eq(int, 2, list_index_of(list, "This is my tail 2 !", my_strcmp)));
-    cr_assert(eq(int, 3, list_index_of(list, "This is my tail 2-1 !", my_strcmp)));
-    cr_assert(eq(int, -1, list_index_of(list, "Invalid Element", my_strcmp)));
+    cr_assert(eq(long, 0, list_index_of(list, "This is my head !", my_strcmp)));
+    cr_assert(eq(long, 1, list_index_of(list, "This is my tail !", my_strcmp)));
+    cr_assert(eq(long, 2, list_index_of(list, "This is my tail 2 !", my_strcmp)));
+    cr_assert(eq(long, 3, list_index_of(list, "This is my tail 2-1 !", my_strcmp)));
+    cr_assert(eq(long, -1, list_index_of(list, "Invalid Element", my_strcmp)));
     list_destroy(list);
 }
